requirements.txt
angles==1.9.12
asgiref==3.3.4
autopep8==1.5.7
base-local-planner==1.16.7
bondpy==1.8.5
catkin==0.7.29
certifi==2020.12.5
chardet==4.0.0
ddynamic-reconfigure-python==0.0.1
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
dj-database-url==0.5.0
Django==3.2
django-filter==2.4.0
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
dynamic-reconfigure==1.6.3
dynamixel-sdk==3.7.31
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.14
gunicorn==20.1.0
idna==2.10
image-geometry==1.13.0
importlib-metadata==4.0.1
interactive-markers==1.11.5
kdl-parser-py==1.13.1
Markdown==3.3.4
message-filters==1.14.10
open-manipulator-with-tb3-tools==1.1.0
phonenumbers==8.12.21
pkg-resources==0.0.0
psycopg2==2.8.6
pycodestyle==2.7.0
python-qt-binding==0.4.3
pytz==2021.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.25.1
resource-retriever==1.12.6
smclib==1.8.5
sqlparse==0.4.1
srdfdom==0.5.2
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
xacro==1.13.9
zipp==3.4.1
